requirements/dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/dev.in -o requirements/dev.txt
annotated-types==0.6.0
    # via pydantic
ansi2html==1.8.0
    # via dash
anyio==3.7.1
    # via
    #   fastapi
    #   httpx
    #   starlette
attrs==23.1.0
    # via
    #   outcome
    #   trio
blinker==1.7.0
    # via flask
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   flask
    #   uvicorn
coverage==7.3.2
    # via pytest-cov
dash==2.14.2
    # via dash-bootstrap-components
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
    # via dash
dash-html-components==2.0.0
    # via dash
dash-table==5.0.0
    # via dash
distlib==0.3.7
    # via virtualenv
fastapi==0.104.1
filelock==3.13.1
    # via virtualenv
flask==3.0.0
    # via dash
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.4
    # via httpx
httpx==0.27.0
identify==2.5.32
    # via pre-commit
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   trio
importlib-metadata==7.0.0
    # via dash
iniconfig==2.0.0
    # via pytest
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.20.1
    # via webargs
mypy==1.7.1
mypy-extensions==1.0.0
    # via mypy
nest-asyncio==1.5.8
    # via dash
nodeenv==1.8.0
    # via pre-commit
numpy==1.26.2
    # via
    #   pandas
    #   pyarrow
outcome==1.3.0.post0
    # via
    #   pytest-trio
    #   trio
packaging==23.2
    # via
    #   marshmallow
    #   plotly
    #   pytest
    #   webargs
pandas==2.1.3
pip-system-certs==4.0
platformdirs==4.1.0
    # via virtualenv
plotly==5.18.0
    # via dash
pluggy==1.3.0
    # via pytest
pre-commit==3.5.0
psycopg2-binary==2.9.9
pyarrow==14.0.1
pydantic==2.5.2
    # via fastapi
pydantic-core==2.14.5
    # via pydantic
pytest==7.4.3
    # via
    #   pytest-cov
    #   pytest-mock
    #   pytest-trio
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-trio==0.8.0
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.1
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via pre-commit
requests==2.31.0
    # via dash
retrying==1.3.4
    # via dash
ruff==0.1.7
setuptools==69.1.1
    # via
    #   dash
    #   nodeenv
six==1.16.0
    # via
    #   python-dateutil
    #   retrying
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   trio
sortedcontainers==2.4.0
    # via trio
sqlalchemy==2.0.27
starlette==0.27.0
    # via fastapi
tenacity==8.2.3
    # via plotly
trio==0.24.0
    # via pytest-trio
typing-extensions==4.8.0
    # via
    #   dash
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2023.3
    # via pandas
urllib3==2.1.0
    # via requests
uvicorn==0.24.0.post1
virtualenv==20.25.0
    # via pre-commit
webargs==8.3.0
werkzeug==3.0.1
    # via
    #   dash
    #   flask
wrapt==1.16.0
    # via pip-system-certs
zipp==3.17.0
    # via importlib-metadata
